microservices/query_processor/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
pymongo>=4.3.3
pymilvus>=2.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pymongo>=2.5.0
google-generativeai>=0.5.0
boto3>=1.30.0
botocore>=1.30